% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPoisson.r
\name{ModelPoisson}
\alias{ModelPoisson}
\title{Automate Logistic Regression Analysis with provided control script and data set}
\usage{
ModelPoisson(
  pathRunType = getwd(),
  user.input = NULL,
  clean = TRUE,
  save.name = "myEnvironment.RData"
)
}
\arguments{
\item{pathRunType}{A directory path to the working folder. Default value is getwd().}

\item{user.input}{File path of the control script to source. Default value is NULL.}

\item{clean}{If clean is TRUE, all folders under the pathRunType directory will be removed. Default value is TRUE.}

\item{save.name}{The name for the saved R object of the modeling results. Default value is "myEnvironment.RData".}
}
\value{
The function does not return any specific object but several objects will be saved as global values during the run.
The model result will be saved as save.name in the directory of pathRunType.
}
\description{
With the provided control script user-input.r and the data set .csv (must be located under the same folder),
ModelPoisson() automate logistic regression analysis on multiple endpoints
including data cleaning, variable selection, model selection and result summary.
}
\examples{
\dontrun{
#after run user-input.r manually
ModelPoisson()

#after setting the path to the working folder
#user-input.r and the data set should be under the working folder already
ModelPoisson(user.input = "user-input.r")

#user-input.r and the data set should be under the working folder already
folder.dir <- getwd()
ModelPoisson(pathRunType = folder.dir,
             user.input = "user-input.r",
             clean = TRUE,
             save.name = "model1.RData")
}

}
