% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportPoisson.R
\name{ReportPoisson}
\alias{ReportPoisson}
\title{Predict the Incidence Rate using New Exposure Data set}
\usage{
ReportPoisson(
  pathRunType = getwd(),
  model.RData = "myEnvironment.RData",
  file.name = NULL
)
}
\arguments{
\item{pathRunType}{A directory path to the working folder. Default value is getwd().}

\item{model.RData}{The saved modeling result from ERModPoisson() or from PredictionPoisson(). Default value is "myEnvironment.RData".}

\item{file.name}{The name for the generated .Rmd file. Default file name is NULL which will save the .Rmd file as "Poisson-Regression-Date&Time.Rmd".}
}
\value{
The function does not return any specific object but several objects will be saved as global values during the run.
A .Rmd file will be saved and it can compile an analysis report (.html) via knitr.
}
\description{
With the provided control script prediction-user-input-sim.r, simulated exposure data set .csv, and the ERModPoisson result .Rdata from ERModPoisson() (must be located under the same folder),
PredictionPoisson() generates incidence rate prediction for new exposures using strudtual model (covariates are ignored).
The prediction curve is only generated when there is a significant relationship between the exposure and the endpoint,
though the summaries of observed incidence rate and the simulated exposure distribution are always generated.
}
\examples{
\dontrun{
# Include prediction of simulated exposures in the report
folder.dir <- getwd()
ModelPoisson(pathRunType = folder.dir,
             user.input = "user-input.r")

PredictionPoisson(pathRunType = folder.dir,
                  prediction.input = "prediction_user_input_sim.R",
                  model.RData = "myEnvironment.RData")

ReportPoisson(model.RData = "myEnvironment.RData")

# No prediction of simulated exposures in the report
#' folder.dir <- getwd()
ModelPoisson(pathRunType = folder.dir,
             user.input = "user-input.r")
ReportPoisson(model.RData = "myEnvironment.RData",
              file.name = "Report_no_pred.Rmd")

}

}
